Try to eliminate the warning

+++ Graphics/GraphicsEngineD3D11/src/ShaderD3D11Impl.cpp
@@ -54,7 +54,7 @@ static const ShaderVersion HLSLValidateShaderVersion(const ShaderVersion& Versio
 
 static const ShaderVersion GetD3D11ShaderModel(D3D_FEATURE_LEVEL d3dDeviceFeatureLevel, const ShaderVersion& HLSLVersion)
 {
-    switch (d3dDeviceFeatureLevel)
+    switch (static_cast<Uint32>(d3dDeviceFeatureLevel))
     {
         // Direct3D11 only supports shader model 5.0 even if the device feature level is
         // above 11.0 (for example, 11.1 or 12.0).
